# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JCETHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with USN and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        usn:
          type: string
          description: University Seat Number
        password:
          type: string
        role:
          type: string
          enum: [student, faculty, admin]
          default: student
      required:
        - usn
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/StudentProfile"
      required:
        - token
        - user
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        usn:
          type: string
        name:
          type: string
        branch:
          type: string
        semester:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarInitials:
          type: string
      required:
        - id
        - usn
        - name
        - branch
        - semester
        - role
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message