# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JCETHub API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with USN and password
 */
export const loginBodyRoleDefault = `student`;

export const LoginBody = zod.object({
  usn: zod.string().describe("University Seat Number"),
  password: zod.string(),
  role: zod.enum(["student", "faculty", "admin"]).default(loginBodyRoleDefault),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.string(),
    usn: zod.string(),
    name: zod.string(),
    branch: zod.string(),
    semester: zod.string(),
    role: zod.string(),
    email: zod.string().optional(),
    phone: zod.string().optional(),
    avatarInitials: zod.string().optional(),
  }),
});

/**
 * @summary Get current student profile
 */
export const GetStudentProfileResponse = zod.object({
  id: zod.string(),
  usn: zod.string(),
  name: zod.string(),
  branch: zod.string(),
  semester: zod.string(),
  role: zod.string(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  avatarInitials: zod.string().optional(),
});

/**
 * @summary Get attendance for logged-in student
 */
export const GetMyAttendanceResponseItem = zod.object({
  subject: zod.string(),
  subjectCode: zod.string(),
  attended: zod.number(),
  total: zod.number(),
  percentage: zod.number(),
  status: zod.enum(["safe", "warning", "low"]),
});
export const GetMyAttendanceResponse = zod.array(GetMyAttendanceResponseItem);

/**
 * @summary Get marks for logged-in student
 */
export const GetMyMarksResponseItem = zod.object({
  subject: zod.string(),
  subjectCode: zod.string(),
  ia1: zod.number(),
  ia2: zod.number(),
  ia3: zod.number(),
  finalMarks: zod.number().optional(),
  maxMarks: zod.number().optional(),
  grade: zod.string().optional(),
});
export const GetMyMarksResponse = zod.array(GetMyMarksResponseItem);

/**
 * @summary Get all notifications
 */
export const GetNotificationsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  message: zod.string(),
  type: zod.enum(["announcement", "alert", "event", "exam"]),
  createdAt: zod.coerce.date(),
  isRead: zod.boolean(),
});
export const GetNotificationsResponse = zod.array(GetNotificationsResponseItem);

/**
 * @summary Get weekly timetable
 */
export const GetTimetableQueryParams = zod.object({
  semester: zod.coerce.string().optional(),
  branch: zod.coerce.string().optional(),
});

export const GetTimetableResponseItem = zod.object({
  day: zod.enum([
    "Monday",
    "Tuesday",
    "Wednesday",
    "Thursday",
    "Friday",
    "Saturday",
  ]),
  period: zod.number(),
  startTime: zod.string(),
  endTime: zod.string(),
  subject: zod.string(),
  subjectCode: zod.string().optional(),
  faculty: zod.string().optional(),
  room: zod.string().optional(),
});
export const GetTimetableResponse = zod.array(GetTimetableResponseItem);

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  overallAttendance: zod.number(),
  cgpa: zod.number(),
  pendingAssignments: zod.number(),
  unreadNotifications: zod.number(),
  currentSemester: zod.string(),
  nextExamDays: zod.number().optional(),
  creditsCompleted: zod.number().optional(),
  creditsTotal: zod.number().optional(),
});

/**
 * @summary Get upcoming college events
 */
export const GetUpcomingEventsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  date: zod.coerce.date(),
  type: zod.enum([
    "exam",
    "holiday",
    "cultural",
    "sports",
    "academic",
    "workshop",
  ]),
  venue: zod.string().optional(),
});
export const GetUpcomingEventsResponse = zod.array(
  GetUpcomingEventsResponseItem,
);